SDK-334 separate python into different files
